perf(ui): copy the rows array once per updateRows batch, not once per matched row
`TracksViewModel.rows` gained a `didSet` in #457 that bumps `rowsVersion`,
which turned the in-place `self.rows[i] = updated` loop in
`updateRows(for:)` into a whole-array copy and a version bump per matched
row. A 50-track rating on a 14k-row Songs list copied 14k rows fifty times.

Both loops now mutate a local copy through one helper and assign back
once, only when a row actually matched, so a batch costs one copy and one
version bump. The rows-version suite gains a multi-row case asserting the
version moves exactly once.

`TrackSortColumn` moves to its own file to keep `TracksViewModel.swift`
under the 500-line limit; the L10n coverage list follows the copy.

Closes #453
